refactor(vmm): move shell and linux_env into mvm-vmm::host
The legacy backends still reached into mvm-runtime for host command
execution. Move the shell helpers and the LinuxEnv dispatch impls down
next to the other shared host substrate so the concrete drivers can run
commands without a back-edge into the orchestration crate.

Re-exported from mvm-runtime::base at the old paths so in-crate callers
and the mvmd contract surface keep resolving.
