+++ crates/mvm-vmm/src/host/mod.rs
@@ -16,10 +16,12 @@ pub mod egress_shared;
 pub mod fc_kernel;
 pub mod host_agent_spawn;
 pub mod hvf_supervisor;
+pub mod linux_env;
 pub mod netd_spawn;
 pub mod observability_target;
 pub mod process_liveness;
 pub mod runtime_meta;
+pub mod shell;
 pub mod snapshot_upper;
 pub mod spec_map;
 pub mod substitution_spawn;
