+++ crates/mvm-vmm/src/host/shell/mock.rs
