+++ crates/mvm-vmm/src/host/shell/mod.rs
